New profile support (#156)
Based on new con4m features, have rewritten a big chunk of chalk to do support "profiles". In the base con4m configuration, key objects are now keyspec objects, and do not take default values. They just specify properties for keys.

Instead, you add "profile" sections, which need to specify which keys the profile will process. When chalking, there are 3 profiles:

The profile used for the actual chalk mark (not used for non-chalking operations).
The profile used for reporting status on individual artifacts during a run.
The profile used for host-level reporting on an individual run.
You then use "outconf" sections to bind profiles to commands. All profile-based outputting happens after all data is collected.

For instance, the outconf section for insert is:

outconf insert  {
  chalk:           "chalking_default"
  artifact_report: "artifact_report_insert_base"
  host_report:     "host_report_insert_base"
}
The three items in strings are profiles defined elsewhere in the config. They're checked for proper usage. BTW, all the core configuration stuff that was in schema.nim is now in definitions.nim which hopefully is a bit better for @drraid, but let me know if you have a better name for it.

Non-chalking commands cannot have the "chalk" field, and if you omit one of the other two items, it's assumed you do not want to report on those things at all.

One big change that results is that plugins need to specify WHEN they collect keys. There are four collection points:

Before any artifacts are searched for-- this is host info collection only, but these fields should generally make sense to stick in either chalk marks or per-artifact reports (or both).
When preparing to chalk-- this is expected to be things that are per-artifact only, so should not be host-level.
When done with chalking one artifact-- this is intended to support per-artifact reporting of information that can only be collected or written, after the chalk has been written into an artifact, so none of these fields are valid in chalk marks.
When done chalking all artifacts-- this is for summary reporting, such as list of unchalked files that codecs recognized, etc.
Additionally, when not inserting, extracted fields will be available for per-artifact reporting.

One big change is that the _ prefix now has an explicit semantic-- it is for data items that CANNOT go into a chalk mark, only host-level reporting. Generally, there's a chalk-friendly version and one that isn't... this is to support collecting metadata about artifacts in their native environment.

That required two changes:

The key that was previously _MAGIC is now just called MAGIC. This will break existing extractions/deletions, so delete any test chalks before updating.
The key that was previously _CHALK_CONFIG has changed... I added a second prefix, '$' (since the con4m lexer already accepts it in IDs), which is for implementation-specific keys.
Most reporting publishes to the "reporting" topic, but you can now add custom reports via a "custom_report" object in the config file.

Major todo items done:

I'd like to be able to have two reporting profiles for non-insert operations... one for if the chalk mark validates, and one if it doesn't. Right now, we'd tend to send the whole mark, but if it validates, we don't need to dupe the artifact metadata, beyond the chalkID and metadata hash seen.
I'd like people to be able to do arbitrary secondary reports tied to whatever sinks they want. Right now, they can install their own profiles, but if they want to tee off data somewhere else, but a more limited set, they'd have to do more work... I want them to write a simple profile.
I'm thinking about adding a profile "inheritance" feature to make them easier to write. It's probably overkill though. (I decided not to do it)
I added back in logic not to collect metadata that won't get used in any reporting.
Need to update the "defaults" command for the new config file schema.
Need to update all the documentation (this is done, minus a polish pass).
More extensive testing of keys, commands, etc. Have validated most things, but not sbom and signing.
Need to go to the example con4m code (sboms, signing, etc) and migrate them.