I didn't have problems getting TLS to work locally once I got DNS (#278)
sorted.  Make sure you're using the right domain in the cert (has to
match what you're browsing to).

However, I did add an optional parameter to the https post sink for
specifying the location of any self-signed cert.  The self-signed-cert
does have to make it to a file, but we configure it via an OpenSSL
call, so people can put the location in a config file, we could use a
tmp file, people can change the env var name, ... (the openssl call
happened in nimutils).

And, in fact, just b/c of testing, the defaultconfig now looks for
TLS_CERT_FILE for the self-signed cert (which is expected to be in PEM
format).

Co-authored-by: Rich Smith <MyNameIsMeerkat@users.noreply.github.com>