Initial procfs plugin implementation (#421)
Initial procfs plugin implementation.  This doesn't come close to exposing all the capabilities of the proc file system. 

Artifact-specific keys only work on `exec` commands, and then only for the exec'd process. The reason this requires a PID to collect, and we currently only collect PIDs on exec() commands.  The core ChalkObj data type has a new field to capture this, though.

Current host-level keys are:
- _OP_TCP_SOCKET_INFO
- _OP_UDP_SOCKET_INFO
- _OP_IPV4_ROUTES
- _OP_IPV6_ROUTES
- _OP_IPV4_INTERFACES
- _OP_IPV6_INTERFACES
- _OP_ARP_TABLE
- _OP_CPU_INFO
- _OP_ALL_PS_INFO

Current per-artifact keys are:
- _PROCESS_PID
- _PROCESS_DETAIL
- _PROCESS_PARENT_PID
- _PROCESS_START_TIME
- _PROCESS_STATE
- _PROCESS_PGID
- _PROCESS_UTIME
- _PROCESS_STIME
- _PROCESS_CHILDREN_UTIME
- _PROCESS_CHILDREN_STIME
- _PROCESS_UMASK
- _PROCESS_UID
- _PROCESS_GID
- _PROCESS_NUM_FD_SIZE
- _PROCESS_GROUPS
- _PROCESS_SECCOMP_STATUS
- _PROCESS_ARGV
- _PROCESS_CWD
- _PROCESS_EXE_PATH
- _PROCESS_COMMAND_NAME
- _PROCESS_FD_INFO
- _PROCESS_MOUNT_INFO