More generic marking of executable source artifacts (#552)
Removed codecShebang and codecPython in favor of codecSource.

This controls whether and how source-based artifacts are marked.

Generally, the marking occurs by sticking the mark in a comment.

Currently, the intent for source marking is to mark content that will
be shipped and run in source code form. While you can mark every
source file, we don't really encourage it. For that reason, by
default, our database only contains reasonably well used scripting
languages.

We also definitely do not recommend marking code while it is in a
repository. Git does that job well, and no tooling exists to help
recalculate every time you make an edit.

Ideally, you might wish to mark both a file and any
dependencies. Currently, with the exception of container images /
containers, Chalk doesn't handle that, as it's significantly difficult
to be particularly precise about what is part of the artifact and what
isn't.
---------

Co-authored-by: indecisivedragon <liming@crashoverride.com>