Nim 2.0 port; Secret service (#627)
Migrated to Nim 2.0, with significant work done to how things get build.

Calling the Secret Service. Currently, there's no auth on the service, which is more an issue for us and less for users, due to the way the crypto is handled. But we should deal with this before releasing (CC @MyNameIsMeerkat).

Closes #521