adding copyright header to all chalk files (#680)
unf none of the tools Ive found so far allow to get
existing autors of the file from either existing copyright
header or from git and all of them just allow to hardcode
an author into copyright header

since copyright belongs to the company and the authors
we can get from git removing that from the header

as for years we can either hardcode specific years
or leave current year in the header. for now using current year
but we can revisit. perhaps we should drop the year altogether
as the copyright notice might not require it at all