Goodbye, `profiles` hello, `templates`
The old notion of `profiles` are now gone, and have been replaced with two things:

* Chalk Mark templates (a `mark_template` section in the config) that specify what chalk-time keys can go into a chalk mark.

* Report templates (a `report_template` section in the config), that are used both to set the current operation's report, and to set up what gets reported in a custom report.

As part of all this, I updated the associated documentation.  The doc system first draft continued to get tweaked here.

It's still pretty messy, but once the renderer in nimutils is improved based on learnings, it'll be good enough for now.