Fixed fd leak regression (#738)
I fixed a regression in file descriptor leaks in the new Elf codec (it
may have only ever been fixed on a branch).

It took me a little too long to find the issue since I had expected it
was fixed. So along the way:

1. I removed the code that uses the nim file walking and replaced it
   with code that provides a simpler API, especially when you may or
   may not want recursion. Nim requires two different library calls,
   and they're slightly different in how they behave.

   That API call went to nimUtils, but the logic in collect.nim now
   basically only runs once; before it had to be repeated 3 times;
   once if the input was a file not a directory, once for dirs when
   recursion was requested, and once for when it wasn't.

2. I got even more defensive on caching file descriptors. When trying
   to isolate the problem, I made a couple plugins not cache on scan, and
   I'm going to leave it that way for a while. It does mean there's
   likely to be more reopening of fds that shouldn't need to happen. But,
   I don't think it'll matter performance wise, and it seems like, as
   we add plugins, we want to be able to see these things faster.

Just to note, when *not* caching, the plugin has opened the stream for
the first time before knowing it needs to be chalked. Generally, if I
pass it to newChalk() I think the caching works right. And certainly
the fd will be re-cached when calling chalk.useFileStream()

But I wanted to be more confident that we were always closing when not
opened via useFileStream(), and now that it's fixed, I don't want to
keep messing w/ things right now, even if it means some unneeded calls
to open().