Initial app inventory doc, and merging of all docs that seem to be nearing completion (#742)
This puts together all of the docs that are nearing completion, in their current state, along with what I believe their associated configs to be. There are still several things that need to be done even on these documents. The configs we should be pushing automatically to the bucket for chalkdust.io.

It does not include in-flight work Liming is doing to the getting started guide and user guides. Those are coming in a separate PR.

@nettrino it also does not contain the outline for the cloud custodian document.

I'll post a list of outstanding items with these docs out of band.