adding tests badge to the repository README
This change adds a badge (svg image) to the repo README
which shows the status of the tests CI pipeline.
Hopefully it will show just that tests are passing
all the time :wink:
The badge also links to the github action page which
shows the history of the tests running on main branch.
Speaking of the main branch, the badge only shows
tests status for the main branch so that if any tests fail
on any PRs, they will not impact the badge status.

Finally this change only runs tests on merges to main
only when any of the source code changes so that we dont
run tests for things like documentation changes.