adding gdb tests for elf encoding (on behalf of Liming)
Liming did most of the work for the gdb stuff so she deserves
the credit!

gdb tests check some basic things about elf encoding by:

* running binary before chalking it and reporting:
  * couple of dozen bytes after the program entrypoint
    as seen in the gdb breakpoint
* chalking the binary
* running same gdb script on the chalked binary and
  comparing the report before chalk and after.
  If all goes well, there should not be any difference
  in the instructions of the entrypoint comparing
  chalked and non-chalked binary.

In the future we can add more checks here as whatever json
gdb script outputs to stdout test will automatically compare.

Speaking of json, some of the bits for looking for valid json
in the os.py/runner.py needed to be changed so that the json
extraction logic can work for both chalk reports as well as other
commands like gdb in this case.