feat(docker): multi-platform docker builds
This is a pretty crude approach which:

* detects if docker build --push --platform=X,Y is used
* if so breaks it into:
* docker build --load for each platform
* docker push for each platform
* docker manifest create
* docker manifest push

As docker manifest is an experimental feature it is guarded so that
if anything fails, docker fallback is used

And it is truly an experimental feature. For example it does not
honor any of the buildx/docker daemon configs for things like insecure
registries so we do our best to fallback in those cases by passing
--insecure flag which is not ideal but does not seem to be any way
around it.

In addition this:

* ensures `CHALK_ID` is the same for all platform builds
* ensures `METADATA_ID` is different for each platform
  by forcing `DOCKER_PLATFORM` to be part of the chalk
* falls-back to docker build on docker push failures
  when buildx is being used (see code comment for context)