fix(docker): wrap_entrypoint honors command args
As chalk exec would parse the full command, if it had any arguments
chalk was not familiar with, chalk would not parse it correctly
and therefore will call command incorrectly.
By adding -- after the command name, chalk ignores rest of the args
and passes them as-is to the command.

For example this now works:

```
ENTRYPOINT ["ls", "-la"]
```

As chalk will end up calling it as:

```
/chalk exec --exec-command-name=ls -- -la
```