refactor: subprocess IO (#91)
Nothing is this PR is directly implemented in Chalk; the implementation was in Nimutils, and this picks up the nimutils version via picking up the newest Con4m.

There's now one incredibly flexible subprocess API, with both a one-shot interface and a builder-style interface. It can do PTYs, capture, callbacks, socket listeners, etc.

The existing subprocess stuff has been all re-written on top of it (we didn't use Nim's built-in stuff because it didn't provide even basic niceties we needed, like being able to capture all of, stdin, stdout and the return code for one process). And it certainly didn't have pty support.

The old API remained for compatibility with chalk though.

The reason for this refactor wasn't the existing subprocess support, it was for future capabilities:

- The file locking is going to get a major upgrade with any parallel chalk processes coordinating via unix sockets.

- We wanted to be able to support capture-replay of Chalk and its use cases (which is now done in: https://github.com/crashappsec/cap10)

- We wanted to be able to SCRIPT sessions too, or partially script, while supporting interaction and capture-replay. Basically a better expect. That's also done in first draft in cap10.