fix(docker): fixing dockerfile path in git context builds
Currently this was not working as expected:

```
chalk docker build -f Dockerfile https://context/repo.git
```

When passing `--file` in docker builds which are using git context,
raw value should be used to find the dockerfile location
vs the normalized `state.dockerFileLoc` variable as it already resolves
it relative to current working directly.
This then allows git context `show()` function to correctly
compute the path within the git context to show the Dockerfile.