fix: supporting shell commands in CMD wrapping
CMD when being a string can be a shell script which directly wrapping
CMD directive breaks. By wrapping it with ENTRYPOINT it allows to
honor any shell scripts

Only edge case is if base image has any custom ENTRYPOINT, that
ENTRYPOINT will break as it will be overwritten. Proper fix will be
coming in next release