refactor: define chalk version only in keyspecs
Before this commit, we had to update the chalk version in two places.
One was in src/configs/base_keyspecs.c4m:

    chalk_version :=   "0.3.1"

And the other in chalk.nimble:

    version       = "0.3.1"

Historically, this duplication has caused the versions to go out of
sync such that a chalk mark contained an incorrect CHALK_VERSION [1].
The tests in chalk now guard against that [2], but it's still better to
define the version in one place.

Make the keyspecs version the single source of truth, setting the
version in the nimble file from there.

This will also help with making a development version of chalk have a
distinguishable version string from a release version of chalk, if
that's something we deem worthwhile.

Also, refactor to remove the macro that was previously used in setting
the version.

Closes: #138

[1] https://github.com/crashappsec/chalk/issues/28
[2] 2f1b55cd70bf, 2023-10-06, "test(lint): adding pre-commit hook for checking chalk version across files"