fix(docker): inspecting base images to find entrypoint
This recursively looks up entrypoints in all base images
from the dockerfile build target. If ENTRYPOINT/CMD/SHELL
is not found in the dockerfile, base image is inspected
to find that. This allows to correctly wrap all builds
as we know exactly how to wrap existing entrypoints.
This in particular will fix the issue for non-shell
ENTRYPOINTs such as for lambda wrappers which expect
CMD to be a python module argument.

In addition this fixes:

* adding docker statements within the target dockerfile section
  boundary. Previously it was always added at the end of Dockerfile
* it honors any --build-arg as passed in via CLI in order
  to eval env vars in the Dockerfile (needed for base image inspection)
* improves error message when docker build fails and fallback is used.
  now exception message will be showed.
* base image config is queried directly from the registry
  via buildx imagetools. This allows to avoid any local docker
  state changes. For example `docker pull alpine --platform=linux/arm64`
  will overwrite local `alpine` image to use arm which can impact
  future builds.
* switching all tests to use trace log level by default.
  this actually found a bug in multi-platform builds as --load
  was not being used by default and so full image metadata
  was not inspected