build: using scratch for base image of public chalk image
As chalk is a static binary now there is a no need to release docker
image based on any distro as we can just release the raw binary image.
This image is not meant to be used directly but is instead meant to be
used in other docker builds such as to easily provision docker build
images used in CI systems.