fix: rename JSon to Json or JSON (#200)
There was some inconsistency in JSON-related identifiers in the
codebase. Nim is partially style insensitive [1], but it's best to make
identifiers consistent to best support e.g. searching for usages with
grep.

The standard capitalization is JSON (for JavaScript Object Notation),
not JSon. So, let's rename:

- JSonNode to JsonNode, which is the upstream style [2]

- CJSonError to CJsonError

- ChalkJSonNode to ChalkJsonNode

- JSon to JSON in comments and exception messages

This is technically a user-facing "fix", because it affects an exception
message.

[1] https://github.com/nim-lang/Nim/blob/v2.0.0/doc/manual.md#identifier-equality
[2] https://github.com/nim-lang/Nim/blob/v2.0.0/lib/pure/json.nim#L188