fix: segfault in chalk backup service
the safeRequest refactor now creates and closes the client
which wasnt correctly refactored and wasnt caught.

there are no tests for this and its not easy to add them
as it will require adding JWTs in the mix. upcoming refactor
for signing keys will refactor this and will make it easier to
test this