fix: honor $CHALK_COMPONENT_CACHE in confload
when loading components on chalk startup, if a component was from a URL
and had any parameters, it would be re-fetched from the URL again
which is not desired as the URL content can change hence causing a
failure.

this was caused by loading parameters before loading component source
which would initialize the component and if missing would re-fetch it
again from URL. By loading source first it removes that possibility.