tests: redefine task for nimble test
By default, running `nimble test` will run every Nim file that begins
with `t` in the `tests` directory [1]. But we've already got a `tests`
directory for all the Python-based tests, and I think it's better to put
the Nim-based tests elsewhere.

Redefine the test task so that it does something useful. Before this
commit, running `nimble test` would do essentially nothing, but print
the misleading:

    Success: All tests passed

[1] https://github.com/nim-lang/nimble/blob/v0.14.2/readme.markdown#tests)