refactor: use fd_cache everywhere
Removing previous chalk FD caching mechanism which stored
file stream on the chalk object which restricted where it can be used.
Now most places use the fd_cache mechanism to interact with FDs
except a few specific places which need to do something custom
such as open FD to TTY which is not cached.