fix: explicitly passing mode for all fd cache yields
If we open by default with fmReadWriteExisting mode during all codec
chalk scans, it leaves all FDs open in write mode which means any
other program using flock on the same file will not work.
For example we scan docker binary to determine if its chalked
and docker seems to use flock and so as chalk would open it in
write mode docker would not be executable in a subprocess.
By explicitly specifying the file mode it allows the caller
to determine how to open the file hence avoiding these type of
conflicts, which are especially pronounced when multiple
chalks run in parallel on the same machine.