fix(docker): docker command name is honored while wrapping docker exe
previously "docker" was only wrapped when "docker" was not present
anywhere in the flags. That is a bad assumption as for example a valid
docker context can be a folder "docker". Right now if the exe name
is "docker" its always appended to the CLI params hence allowing
chalk to correctly validate CLI args.