refactor(techStackGeneric): ignore .git directory (#253)
A significant fraction of the tech stack detection execution time was
often from scanning files in any `.git` directory. Prevent that.

Clearly, the speedup here depends on the relative size of the `.git`
directory versus the rest of the data. But at least on one machine,
this commit is a 2x speedup for running in the chalk repo:

    chalk insert --use-tech-stack-detection ./foo

We should eventually improve the directory walking here, and consider
ignoring other directories and files, but let's just do this for now.

Refs: https://github.com/crashappsec/chalk/issues/249