ci: use ubuntu-22.04, not ubuntu-latest
The ubuntu-latest label currently uses Ubuntu 22.04 [1], but will likely
use Ubuntu 24.04 later this year. Pin the Ubuntu version so that our
upgrading is explicit.

The ubuntu-latest label started using Ubuntu 22.04 at the end of 2022,
but the rollout was advertised to take 8 weeks [2]. This commit prevents
possible complications due to the runner image changing from one
workflow run to the next, even with the same repo state.

See some example differences between the installed and default versions
of tools on different images [3][4].

Refs: https://github.com/crashappsec/chalk/issues/142

[1] https://github.com/actions/runner-images/blob/07d23cf65fdc/README.md#available-images
[2] https://github.blog/changelog/2022-11-09-github-actions-ubuntu-latest-workflows-will-use-ubuntu-22-04/
[3] https://github.com/actions/runner-images/blob/07d23cf65fdc/images/ubuntu/Ubuntu2004-Readme.md
[4] https://github.com/actions/runner-images/blob/07d23cf65fdc/images/ubuntu/Ubuntu2204-Readme.md