refactor: splitting docker inspection logic
Previously a lot of the logic was in docker_base which was very hard to
follow. In addition there were some tidbits which were defined in other
files such as codecDocker which made things hard to reason about.
Everything was consolidated now and split in submodules in docker/
folder. Important modules:

* inspect - runs docker inspect command
* collect - map docker metadata to chalk object
* scan - find new chalk objects to collect info for

All other modules use these primary functions to implement rest of the
functionality. As this refactor mostly focused on metadata refactor,
other functionality was left as-is but should be refactored in the
future. There is still a lot of things going on in docker/base.nim