chalk_common, manifest: DockerPlatform: convert from tuple to object
When defining types in type sections, it's generally better to use
`object` over `tuple`.

This generally improves readability and can improve type safety. Tuples
are structurally typed [1]:

    Different tuple-types are equivalent if they specify the same fields
    of the same type in the same order. The names of the fields also
    have to be the same.

    [...]

    Objects provide many features that tuples do not. Objects provide
    inheritance and the ability to hide fields from other modules.

    [...]

    In contrast to tuples, different object types are never equivalent,
    they are nominal types whereas tuples are structural.

which introduces the possibility to e.g. incorrectly compare two tuples
with the same fields.

Try to use tuple only when it aids readability for returning multiple
values, or unpacking like:

    let (a, b) = foo()

Which I'd say isn't the case here.

Aside: it's possible to declare tuples in type sections without `[` and
`]`, and commas:

    type Foo = tuple
      a: int
      b: int

[1] https://nim-lang.org/2.0.0/manual.html#types-tuples-and-object-types