ci: add testament-based unit testing (#227)
Chalk is a binary, not a library, so the most important thing to test
is the binary itself. We currently use pytest for those integration
tests.

However, unit tests are still valuable. Allow the user to run unit
tests with:

    nimble test

or to use the same Docker setup as the Python-based tests:

    make unit-tests

and add a workflow to run the latter in CI.

By default, running `nimble test` will run every Nim file that begins
with `t` in the `tests` directory [1]. This commit redefines the test
task to use testament [2] to run all Nim files in the `tests/unit`
directory, and moves the existing Python-based tests to
`tests/functional`. Before this commit, running `nimble test` would do
essentially nothing, but output the misleading:

    Success: All tests passed

As a starting point, move the previously in-file unit tests for semver
parsing to the new directory. Also alter them to use doAssert [3] as
upstream recommends [4], and test at compile time via the
`static: main()` pattern. These tests themselves are not important -
the main point of this commit is to:

- Give us something to iterate on as appropriate.

- Make it easy to add unit tests for new code, aiding development and
  code review.

- Make it easy to add unit tests for old code when valuable.

Closes: https://github.com/crashappsec/chalk/issues/226

[1] https://github.com/nim-lang/nimble/blob/v0.14.2/readme.markdown#tests
[2] https://nim-lang.org/2.0.0/testament.html
[3] https://github.com/nim-lang/Nim/blob/version-2-0/lib/std/assertions.nim#L68-L72
[4] https://nim-lang.org/2.0.0/contributing.html#writing-tests-stdlib

Co-authored-by: liming <liming@crashoverride.com>