fix(cosign): push attestation to each registry
As docker images can be pushed to multiple registries at once,
cosign needs to push to all of those registries as well so that
it will be able to validate signatures from any of them in the future.
Previously only the first tag attestation was pushed to.

As cosign generates signature by the full registry name, when pushing
to multiple registries, each registry ends up with a unique signature.
As a result changing `_SIGNATURE` field to `_SIGNATURES` to accommodate
that.