fix(docker): wrapping image when base image defines USER
Previously we were only detecting USER from the Dockerfile
which meant that it was possible for build to fail if base image
would specify USER which was not compatible with copying chalk mark
or binary into the image.

Now we inspect the base image USER and then make the files available
in dockerfile appropriately.

Created wrap.nim as base.nim was getting a lit more long with all the
wrapping logic.