feat(docker): multi-platform builds auto download chalks
Previously the user had to explicitly configure chalk with every path
where it should look for other architecture chalk binaries for docker
multi-platform builds. That was not very user-friendly as it required
additional setup on the host to be able to fully do multi-platform
builds.

Now chalk will look for other arch binaries in this precedence:

* Honor user-set paths as per `docker.arch_binary_locations`
* Look for binaries in `docker.arch_binary_locations_path`
* Download the release binary for the same chalk version

This still allows for user-flexibility such as during testing
however it makes default behavior for multi-platform builds
friction-less.

To facilitate that, `chalk dump` and `chalk load` where adjusted
to allow one chalk to fully copy its user-config to another chalk
binary. The flow it uses is roughly:

```
chalk dump all | chalk load --replace --all -
```

That automatically copies all user configs which includes:

* user config
* component cache
* component parameters
* any memoized values from the user-config

As the config is validated during `chalk load`, this should ensure
compatibility between chalk versions in case running chalk is
not compatible with binaries already found on disk.