fix: supporting docker platform variants
its slightly complicated as some variants are normalized
so for example linux/amd64/v8 is normalized to linux/amd64
and so when we look for base images we need to compare platforms
by their normalized form

some platforms however are not normalized correctly and docker
has ambiguous behavior in which case chalk fails to wrap them.
for example linux/arm/v8 docker pulls v7 image but keeps using
v8 in the TARGETPLATFORM build arg :shrug:. As chalk does not
support those other platforms this should not impact any
legitimate usage.