test: testing semgrep/syft with/without docker
needed to add *_prefer_docker config to allow to set docker precedence so that
in tests we can force to use docker even if the tool is already installed on
the system. Otherwise would of required to run tests in particular order which
is a bad idea in the long term.