feat: external tools if not installed run via docker (#314)
* feat: external tools if not installed run via docker

previously chalk would always try to install the external tool on the
host system which in some cases required some system dependencies like
`pipx` or `python3` to be available for `semgrep`. Right now the
precedence is:

1. use installed tool
2. use docker (if present)
3. install tool

This should minimize any external system dependencies and should allow
to bundle any tool configs in a container which can be customized via
the config:

```
tool.syft.syft_container = "mysyft:latest"
```

In addition some logging was improved in tool runtime to help in
debugging what tool ran and what keys it generated.

Also `semgrep ci` does not allow to specify target for scanning which
means that it was always scanning `cwd()` which is incorrect for example
in case of docker builds when context is outside of `cwd()`.
Now `semgrep scan` is used which allows to specify target for scanning.

* docs: english

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>

* test: testing semgrep/syft with/without docker

needed to add *_prefer_docker config to allow to set docker precedence so that
in tests we can force to use docker even if the tool is already installed on
the system. Otherwise would of required to run tests in particular order which
is a bad idea in the long term.

---------

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>