fix(git): commit id is correctly determined for annotated symbolic ref tags (#347)
`.git/HEAD` can be a symbolic ref to a branch/tag or directly pointer to
a commit. When doing `git checkout <branch>` it normally sets it to
symbolic ref however when doing `git checkout <tag>` it usually sets it
directly to the tag's commit. However it is possible to update it via
`git symbolic-ref HEAD <tag>`. In that case `HEAD` will point to
annotated tag and as such we chalk simply opens the ref of that symbolic
ref it is not guaranteed to be for a commit id. Now chalk loads the tag
object and if it is an annotated tag, resolves commit id to be from the
tags object pointer.