fix: refetching lightweight tags from origin (#349)
* fix: refetching lightweight tags from origin

If repo is checked out incorrectly locally, a tag can be lightweight
but annotated in origin. To guarantee chalk sees up-to-date tags,
chalk refetches lightweight tags to ensure it reports correct metadata
about the "latest" tag.

This was noticed via checkout github action which if triggered from a
tag build it looses any tag annotations.