fix(docker): fixing --metadata-file support
In older docker versions, as `docker build` is not an alias
to `docker buildx build`, `--metadata-file` flag is not supported.
It only works when explicitly using `docker buildx build` in which case
0.6.0 is required which added that flag. Otherwise docker version
needs to be >=22 which aliased `build` to `buildx build` as well as
have supported buildx version.

No easy way of testing it in CI but tested locally various combination
of versions.